feat: update pull to refresh functionality on wallet and home tabs (#6255)
### Description

As the title:
1. on the Wallet tab, pull to refresh refreshes balances. the loader is
also displayed for any background refreshes.
2. on the Home tab, pull to refresh refreshes only the transaction
items. the loader is not displayed for background refreshes.

### Test plan



https://github.com/user-attachments/assets/b0a56779-e02c-4f2f-b5e9-08b2bd6dfd9a



### Related issues

- Fixes RET-1254

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)
